Revert "Documentation, Page structure, filler pages, find events by day, etc."
