--- azenda/app/useraccount/[username]/add_repeating_events/page.js
@@ -1,1 +1,0 @@
-'use client'
