+++ azenda/app/useraccount/[username]/edit/[event_id]/page.js
@@ -4,14 +4,14 @@
 import { useEffect, useState } from "react";
 import { useParams, useRouter, useSearchParams } from "next/navigation";
 
-async function addEvent(event_name, allows_concurrent_events, event_user, start_time_year, start_time_month, start_time_day, start_time_hour, start_time_minute, duration_mins)
+async function modifyEvent(id, data)
 {
-    const res = await fetch("http://127.0.0.1:8000/api/createEvents/", {
-        method: "POST",
+    const res = await fetch(`http://localhost:8000/api/modify_event/${id}/`, {
+        method: "PUT",
         headers: {
           "Content-Type": "application/json",
         },
-        body: JSON.stringify({event_name, allows_concurrent_events, event_user, start_time_year, start_time_month, start_time_day, start_time_hour, start_time_minute, duration_mins}),
+        body: JSON.stringify(data),
       });
     
       if (!res.ok) {
@@ -21,19 +21,27 @@ async function addEvent(event_name, allows_concurrent_events, event_user, start_
       return res.json();
     
 }
+async function getEvent(id)
+{
+    const res = await fetch(`http://localhost:8000/api/modify_event/${id}/`, {
+        method: "GET",
+    });
+
+    if (!res.ok) {
+        throw new Error("Failed to create data");
+      }
+    
+      return res.json();
+}
 
 const Page = () => {
     const router = useRouter();
-    const [formData, setFormData] = useState({ event_name: "",  start_time_hour: "", start_time_minute: "", duration_mins: ""});
+    const [formData, setFormData] = useState({event_name: "", start_time: "", duration_mins: ""});
     const [isLoading, setIsLoading] = useState(false);
     const [error, setError] = useState(null);
     const params = useParams();
-    
-    //gets dynamic parameters
+    const eventid = params.event_id;
     const user_name = params.username;
-    const full_date = params.custom_date;
-    const [date_month, date_day, date_year] = full_date.split("-");
-
     /**
      * Handles the form submission.
      * @param {Event} event The form submission event.
@@ -41,16 +49,16 @@ const Page = () => {
     const onFinish = (event) => {
       event.preventDefault();
       setIsLoading(true);
-      
-      addEvent(formData.event_name, false, user_name, date_year, date_month, date_day, formData.start_time_hour, formData.start_time_minute, formData.duration_mins)
+      modifyEvent(eventid, formData)
         .then(() => {
           // Navigate to the main page with a query parameter indicating success
           //router.replace("/?action=add");
-          router.replace(`../${full_date}`);
+          router.replace(`/useraccount/${user_name}`);
         })
-        .catch(() => {
-          setError("An error occurred");
-          setIsLoading(false);
+        .catch((error) => {
+            console.error("Error details:", error);
+            setError("An error occurred");
+            setIsLoading(false);
         });
     };
   
@@ -58,7 +66,18 @@ const Page = () => {
     useEffect(() => {
       return () => setIsLoading(false);
     }, []);
-  
+
+    useEffect(() => {
+        const fetchData = async () => {
+          try {
+            const data = await getEvent(eventid);
+            setFormData({event_name: data.event_name, start_time: data.start_time, duration_mins: data.duration_mins});
+          } catch (error) {
+            setError(error.message);
+          }
+        };
+        fetchData();
+      }, [eventid]);
     return (
       <form onSubmit={onFinish}>
         <div className="form-item">
@@ -73,29 +92,17 @@ const Page = () => {
           />
         </div>
         <div className="form-item">
-          <label htmlFor="start_time_hour">Hour of Event</label>
+          <label htmlFor="start_time">Time of Event</label>
           <input
             required
-            name="start_time_hour"
-            type="number"
-            value={formData.start_time_hour}
+            name="start_time_year"
+            value={formData.start_time}
             onChange={(event) =>
-              setFormData({ ...formData, start_time_hour: event.target.value })
+              setFormData({ ...formData, start_time: event.target.value })
             }
           />
         </div>
-        <div className="form-item">
-          <label htmlFor="start_time_minute">Minute of Event</label>
-          <input
-            required
-            name="start_time_minute"
-            type="number"
-            value={formData.start_time_minute}
-            onChange={(event) =>
-              setFormData({ ...formData, start_time_minute: event.target.value })
-            }
-          />
-        </div>
+
         <div className="form-item">
           <label htmlFor="duration_mins">Duration of Event(in minutes)</label>
           <input
