--- azenda/app/useraccount/[username]/optimized_tasks/page.js
@@ -1,1 +1,0 @@
-'use client'
