+++ azenda/backend/accounts/views.py
@@ -10,13 +10,8 @@ from datetime import datetime
 from rest_framework.parsers import JSONParser 
 from django.views.decorators.csrf import csrf_exempt
 
-
-
-#Account Management functions
-#Mainly pertaining to user accounts
-
 @api_view(['POST'])
-def login(request): #login a user
+def login(request):
     username = request.data.get('username')
     password = request.data.get('password')
 
@@ -27,17 +22,12 @@ def login(request): #login a user
     else:
         return Response({"error": "Invalid credentials"}, status=status.HTTP_401_UNAUTHORIZED)
     '''
-
-    #check if the username the user put in even exists, if it doesn't go to except
     
     try:
-        account = Account.objects.get(username=username)
+        account = Account.objects.get(username=username)#check if the username the user put in even exists
         
-        #Correct username and password
-        if account.password == password: 
+        if account.password == password:
             return Response({"message": "Username and password match"}, status=status.HTTP_200_OK)
-        
-        #username exists, but incorrect password
         else:
             return Response({"message": "Password does not match"}, status=status.HTTP_401_UNAUTHORIZED)
     
@@ -45,19 +35,15 @@ def login(request): #login a user
         return Response({"message": "Username not found"}, status.HTTP_404_NOT_FOUND)
     
 
-@api_view(['POST']) #create an account and handle exceptions
+@api_view(['POST']) 
 def createaccount(request):
-
-    #user parameters i.e. username, password
     _username = request.data.get('username')
     _password = request.data.get('password')
     _sleep_start = request.data.get('sleep_start')
     _sleep_end = request.data.get('sleep_end')
-    
     try:
         accountcheck = Account.objects.get(username=_username)
         return Response({"message": "Account already exists"}, status=status.HTTP_200_OK)
-    
     except Account.DoesNotExist:
         try:
             new_account = Account(username=_username, password=_password, sleep_start=_sleep_start, sleep_end=_sleep_end)
@@ -70,40 +56,23 @@ def createaccount(request):
 
 
 
-
-#Event CRUD Operations
-#Mainly adding, removing, editing, and retrieving events
-
-@api_view(['GET']) #get events for a specific user
+@api_view(['GET'])
 def retrieveUserEvents(request):
     alluserevents = Event.objects.all()
     path = request.query_params.get('username', None)
-    alluserevents = alluserevents.filter(event_user=path).order_by('start_time')
+    alluserevents = alluserevents.filter(event_user=path)
     if path is not None:
         event_serializer = EventSerializer(alluserevents, many=True)
     return JsonResponse(event_serializer.data, safe=False)
 
-@api_view(['GET']) #get events for a specific user from a specific day
-def retrieve_Events_from_Day(request):
-    alluserevents = Event.objects.all()
-    path = request.query_params.get('username', None)
-    date_year = request.query_params.get('date_year', None)
-    date_month = request.query_params.get('date_month', None)
-    date_day = request.query_params.get('date_day', None)
-    alluserevents = alluserevents.filter(event_user=path, start_time__year = date_year, start_time__month = date_month, start_time__day = date_day).order_by('start_time')
-    if path is not None:
-        event_serializer = EventSerializer(alluserevents, many=True)
-    return JsonResponse(event_serializer.data, safe=False)
-
-
-@api_view(['POST']) #Add New Event
+@api_view(['POST'])
 def createEvent(request):
     _event_name = request.data.get('event_name')
     _allows_concurrent_events = request.data.get('allows_concurrent_events')
 
 
     event_user_id = request.data.get('event_user')
-    _event_user = get_object_or_404(Account, username=event_user_id) #foreign key for user
+    _event_user = get_object_or_404(Account, username=event_user_id)
 
 
     _start_time_year = request.data.get('start_time_year')
@@ -120,22 +89,21 @@ def createEvent(request):
                  int(_start_time_hour), 
                  int(_start_time_minute))
     
-    try:#whole bunch of error handling
+    try:
         new_event = Event(event_name = _event_name, 
                           allows_concurrent_events = _allows_concurrent_events, 
                           event_user = _event_user, 
                           start_time = d, 
                           duration_mins = int(_duration_mins))
         new_event.save()
-        return Response({"message": "Event Created"}, status=status.HTTP_201_CREATED)#correct functioning case, need an additional case for handling event conflicts
+        return Response({"message": "Event Created"}, status=status.HTTP_201_CREATED)
     except IntegrityError:
         return Response({"message": "Failed to create account due to database error."}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
     except Exception as e:
         return Response({"message": f"An unexpected error occurred: {str(e)}"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
     
-
 @csrf_exempt
-def modify_event(request, event_id): #Edit an event
+def modify_event(request, event_id):
     try: 
         searchevent = Event.objects.get(id=event_id) 
     except Event.DoesNotExist:
