+++ azenda/backend/backend/urls.py
@@ -16,7 +16,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from accounts.views import login, createaccount, retrieveUserEvents, createEvent, modify_event, retrieve_Events_from_Day  # Import the login view
+from accounts.views import login, createaccount, retrieveUserEvents, createEvent, modify_event  # Import the login view
 from accounts.routers import router
 
 
@@ -24,9 +24,9 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/login/', login, name='login'),  # Add the login API endpoint
     path('api/createaccount/', createaccount, name='createaccount'), # account creation endpoint
-    path('api/retrieveUserEvents/', retrieveUserEvents, name='retrieveUserEvents'),   #grab a user's events by ids
-    path('api/createEvents/', createEvent, name='createEvent'), #make an event for a specific user
-    path('api/modify_event/<int:event_id>/', modify_event, name='modify_event'), #path for modifying events, has dynamic parameter for the id of the event to be modified
+    path('api/retrieveUserEvents/', retrieveUserEvents, name='retrieveUserEvents'),
+    path('api/createEvents/', createEvent, name='createEvent'),
+    path('api/modify_event/<int:event_id>/', modify_event, name='modify_event'),
     path('api/', include((router.urls, 'core_api'), namespace='core_api')),
-    path('api/retrieve_Events_from_Day/', retrieve_Events_from_Day, name='retrieve_Events_from_Day')
+
 ]
